Move npm only test into separate class and ignore it for bower mode
Clean imports in ITs
