+++ flow-tests/test-embedding/embedding-test-assets/src/test/java/com/vaadin/flow/webcomponent/EmbeddedWebComponentIT.java
@@ -17,8 +17,6 @@ package com.vaadin.flow.webcomponent;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
 
 import com.vaadin.flow.testutil.ChromeBrowserTest;
 import com.vaadin.testbench.TestBenchElement;
