+++ flow-tests/test-embedding/embedding-test-assets/src/test/java/com/vaadin/flow/webcomponent/PreserveOnRefreshIT.java
@@ -15,14 +15,11 @@
  */
 package com.vaadin.flow.webcomponent;
 
-import java.util.List;
-
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
-import org.openqa.selenium.WebElement;
 
 import com.vaadin.flow.testutil.ChromeBrowserTest;
 import com.vaadin.testbench.TestBenchElement;
