+++ flow-tests/test-embedding/test-embedding-production-mode-compatibility/pom.xml
@@ -79,6 +79,10 @@
                 <artifactId>maven-failsafe-plugin</artifactId>
                 <configuration>
                     <dependenciesToScan>com.vaadin:flow-embedding-test-assets</dependenciesToScan>
+                    <excludes>
+                        <!-- Remove once #5732 is fixed -->
+                        <exclude>**/NpmOnlyIndexIT.java</exclude>
+                    </excludes>
                 </configuration>
             </plugin>
